+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to `laravel-ban` will be documented in this file.
 
+## 2.0.0 - 2017-03-06
+
+### Changed
+
+- Contract `CanBeBanned` renamed to `HasBans`
+- Trait `CanBeBanned` renamed to `HasBans`
+
 ## 1.0.0 - 2017-03-05
 
 - Initial release
+
+[2.0.0]: https://github.com/cybercog/laravel-ban/compare/1.0.0...2.0.0